# docs/public_api_reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "beetkeeper",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": ""
  },
  "paths": {
...
    "/api/import": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Start Import",
        "description": "Enqueue an import of the given paths and return the created (PENDING) job.\n\nSet `quiet=true` to import non-interactively (the `beet import -q` equivalent): no decision prompts.\nThe optional per-job settings (`quiet`, `logpath`, `group_albums`, `flat`, `set_fields`) default to the\ncorresponding beets config values when left unspecified.",
        "operationId": "start_import_api_import_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "List Imports",
        "description": "List one page of known import jobs, newest first (so page 1 shows the most recent submissions).",
        "operationId": "list_imports_api_import_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "The page number, 1-indexed.",
              "default": 1,
              "title": "Page"
            },
            "description": "The page number, 1-indexed."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum results per page.",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Maximum results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportJob"
                  },
                  "title": "Response List Imports Api Import Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/clean_slate/preview": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Preview Clean Slate",
        "description": "Dry-run a clean-slate import without touching anything (see `POST /api/import/clean_slate`).\n\nReports what the removal would delete (library files inside the beets directory, album art), what it\nwould leave alone, the flexible attributes that would be lost and those re-applied to the fresh import\n(`fields_preserved`: the entry's values for the keys of the `downloader_hook.beet_field_to_dl_search_field`\nconfig), what the source folder holds, plus the blocking `errors`, non-blocking `warnings`, and whether the\nfewer-files opt-in is needed. 404 for an unknown beets id.",
        "operationId": "preview_clean_slate_api_import_clean_slate_preview_get",
        "parameters": [
          {
            "name": "beets_album_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The beets library id of an album. Mutually exclusive with `beets_item_id`.",
              "title": "Beets Album Id"
            },
            "description": "The beets library id of an album. Mutually exclusive with `beets_item_id`."
          },
          {
            "name": "beets_item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The beets library id of an item (track). Mutually exclusive with `beets_album_id`.",
              "title": "Beets Item Id"
            },
            "description": "The beets library id of an item (track). Mutually exclusive with `beets_album_id`."
          },
          {
            "name": "source_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "The entry's raw (pre-import) source folder, under beetkeeper's `downloads_path`; for a standalone track, the audio file itself or a folder holding just that file.",
              "title": "Source Path"
            },
            "description": "The entry's raw (pre-import) source folder, under beetkeeper's `downloads_path`; for a standalone track, the audio file itself or a folder holding just that file."
          },
          {
            "name": "allow_fewer_files",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Proceed even when the source holds fewer audio files than the entry currently has on disk (the clean slate would then delete files the source cannot replace). Off by default.",
              "default": false,
              "title": "Allow Fewer Files"
            },
            "description": "Proceed even when the source holds fewer audio files than the entry currently has on disk (the clean slate would then delete files the source cannot replace). Off by default."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanSlatePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/clean_slate": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Start Clean Slate",
        "description": "Enqueue a clean-slate import: remove the named library entry, then import its source folder afresh.\n\nThe preview runs first: a blocking error is a 422, a source with fewer audio files than the entry has on\ndisk is a 409 until `allow_fewer_files` is set, and an unknown beets id is a 404. The job is then\ntracked like any other import (poll it, answer its decisions, abort it via the routes below); the worker\nre-runs the preview as its guard right before removing anything.",
        "operationId": "start_clean_slate_api_import_clean_slate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanSlateSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/find_missing_source_path": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Find Missing Source Path",
        "description": "Ask the configured downloader client where a library album/track was originally downloaded to.\n\nOnly useful for entries imported outside a beetkeeper context (so no source path was recorded): the\nrecovered pre-import folder is what a clean-slate import of the entry needs. The entry's fields named in\nthe beets config's `beetkeeper.downloader_hook.beet_field_to_dl_search_field` become the search\nrequest's query params (see the `search-missing-source-path` webhook). A match is stored as the entry's\n*inferred* source path (replacing any earlier inference), which the search page then shows labelled as\ninferred \u2014 separate from source paths recorded from the beetkeeper plugin's events. 409 when no\ndownloader API is configured; 404 for an unknown beets id.",
        "operationId": "find_missing_source_path_api_import_find_missing_source_path_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindMissingSourcePathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindMissingSourcePathResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/{job_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Import",
        "description": "Return a single import job (poll this for status / the pending decision).",
        "operationId": "get_import_api_import__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/{job_id}/decision": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Decide Import",
        "description": "Answer the decision an import is parked on; 409 if it isn't awaiting one.",
        "operationId": "decide_import_api_import__job_id__decision_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/{job_id}/abort": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Abort Import",
        "description": "Request cooperative cancellation of an in-flight import.",
        "operationId": "abort_import_api_import__job_id__abort_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "CleanSlatePreview": {
        "properties": {
          "subject": {
            "type": "string",
            "enum": [
              "album",
              "track"
            ],
            "title": "Subject"
          },
          "beets_id": {
            "type": "integer",
            "title": "Beets Id"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "`albumartist - album` for an album, `artist - title` for a standalone track."
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count",
            "description": "Library rows the entry holds (1 for a standalone track)."
          },
          "files_present": {
            "type": "integer",
            "title": "Files Present",
            "description": "How many of those rows still have their file on disk."
          },
          "missing_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Paths",
            "description": "Library paths whose file is gone."
          },
          "expected_tracks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Tracks",
            "description": "Track total the album's tags claim (beets' `albumtotal`), when known."
          },
          "files_to_delete": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files To Delete",
            "description": "Library files the removal deletes (inside the beets directory)."
          },
          "files_kept_outside_library": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files Kept Outside Library",
            "description": "Entry files left on disk because they live outside the beets directory."
          },
          "art_to_delete": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Art To Delete",
            "description": "The album art file the removal deletes, if any."
          },
          "flexible_attributes_lost": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flexible Attributes Lost",
            "description": "Flexible attribute names on the entry that a clean slate does not keep (`fields_preserved` aside)."
          },
          "fields_preserved": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Fields Preserved",
            "description": "The entry's flexible attributes named as keys of the `downloader_hook.beet_field_to_dl_search_field` config, with their values: re-applied to the fresh import via beets' `--set`."
          },
          "source_path": {
            "type": "string",
            "title": "Source Path"
          },
          "source_audio_files": {
            "type": "integer",
            "title": "Source Audio Files",
            "description": "Readable audio files found under the source path.",
            "default": 0
          },
          "source_album_groups": {
            "type": "integer",
            "title": "Source Album Groups",
            "description": "Album folders beets would import from the source.",
            "default": 0
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "needs_confirmation": {
            "type": "boolean",
            "title": "Needs Confirmation",
            "description": "The source has fewer audio files than the entry has on disk and `allow_fewer_files` was not set.",
            "default": false
          },
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "description": "Whether the clean slate may run as previewed (no blocking error and no outstanding confirmation).",
            "readOnly": true
          },
          "files_missing": {
            "type": "integer",
            "title": "Files Missing",
            "description": "How many of the entry's rows point at a file that no longer exists.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "subject",
          "beets_id",
          "label",
          "item_count",
          "files_present",
          "source_path",
          "ok",
          "files_missing"
        ],
        "title": "CleanSlatePreview",
        "description": "The dry run of a clean-slate import: what removing the library entry and importing its source would do.\n\nProduced read-only by `core.clean_slate.preview` for the API/UI preview, and again by the worker as the\nguard before it removes anything. `errors` lists the conditions that block the job outright; `warnings`\nare worth a look but do not block. `needs_confirmation` is set when the source folder holds fewer audio\nfiles than the entry currently has on disk \u2014 proceeding would delete files the source cannot replace,\nso the caller must opt in explicitly (`allow_fewer_files`)."
      },
...
      "FindMissingSourcePathRequest": {
        "properties": {
          "beets_album_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beets Album Id",
            "description": "The beets library id of an album. Mutually exclusive with `beets_item_id`."
          },
          "beets_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beets Item Id",
            "description": "The beets library id of an item (track). Mutually exclusive with `beets_album_id`."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FindMissingSourcePathRequest",
        "description": "Body of `POST /api/import/find_missing_source_path`: the entry to ask the downloader about."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ImportJobStatus"
          },
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paths"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "pending_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision_submitted": {
            "type": "boolean",
            "title": "Decision Submitted",
            "default": false
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "quiet": {
            "type": "boolean",
            "title": "Quiet",
            "default": false
          },
          "logpath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logpath"
          },
          "group_albums": {
            "type": "boolean",
            "title": "Group Albums",
            "default": false
          },
          "flat": {
            "type": "boolean",
            "title": "Flat",
            "default": false
          },
          "set_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Set Fields"
          },
          "clean_slate_album_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clean Slate Album Id"
          },
          "clean_slate_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clean Slate Item Id"
          },
          "clean_slate_allow_fewer_files": {
            "type": "boolean",
            "title": "Clean Slate Allow Fewer Files",
            "default": false
          },
          "is_clean_slate": {
            "type": "boolean",
            "title": "Is Clean Slate",
            "description": "Whether this job removes an existing library entry before importing its source afresh.",
            "readOnly": true
          },
          "source_label": {
            "type": "string",
            "title": "Source Label",
            "description": "Display name for the import source: the path basename(s), prefixed for a clean-slate import.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "paths",
          "created_at",
          "is_clean_slate",
          "source_label"
        ],
        "title": "ImportJob",
        "description": "API-facing view of a single import job (the persisted record is `db.models.ImportJobRecord`)."
      },
      "ImportJobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "awaiting_decision",
          "completed",
          "failed",
          "aborted"
        ],
        "title": "ImportJobStatus",
        "description": "Lifecycle states of an import job."
      },
      "ImportSubmitRequest": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Paths",
            "description": "One or more paths (files/dirs) to import."
          },
          "quiet": {
            "type": "boolean",
            "title": "Quiet",
            "description": "Run non-interactively (like `beet import -q`): never prompt. Strong matches are applied automatically; anything else falls back to beets' `quiet_fallback` config (skip by default)."
          },
          "logpath": {
            "anyOf": [
              {
                "type": "string",
                "format": "path"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logpath",
            "description": "Corresponds to the `-l` option when running an import command, or config's `import.log` option. See https://beets.readthedocs.io/en/stable/reference/cli.html#import"
          },
          "group_albums": {
            "type": "boolean",
            "title": "Group Albums",
            "description": "See https://beets.readthedocs.io/en/stable/reference/config.html#group-albums"
          },
          "flat": {
            "type": "boolean",
            "title": "Flat",
            "description": "Corresponds to the `--flat` flag when running an import command. See https://beets.readthedocs.io/en/stable/reference/cli.html#import"
          },
          "set_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Set Fields",
            "description": "Corresponds to the `--set field=value` option for beets' `import` CLI command. Can set multiple key-value pairs."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paths"
        ],
        "title": "ImportSubmitRequest",
        "description": "Body for starting an import: the filesystem path(s) beets should import. All default values come from the\n`beets.config` values (cached) if left unspecified."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}